rename preview to preview-styles for consistency
