+++ packages/edit-site/src/components/global-styles/style-variations-container.js
@@ -10,7 +10,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import StylesPreview from './preview';
+import PreviewStyles from './preview-styles';
 import Variation from './variations/variation';
 
 export default function StyleVariationsContainer() {
@@ -43,7 +43,7 @@ export default function StyleVariationsContainer() {
 			{ withEmptyVariation.map( ( variation, index ) => (
 				<Variation key={ index } variation={ variation }>
 					{ ( isFocused ) => (
-						<StylesPreview
+						<PreviewStyles
 							label={ variation?.title }
 							withHoverView
 							isFocused={ isFocused }
